Add PlaybackSpeedMenu
